Recap

+++ 0x1A-hash_tables/2-key_index.c
@@ -9,9 +9,8 @@
  */
 unsigned long int key_index(const unsigned char *key, unsigned long int size)
 {
-    unsigned long int idx;
+	uint idx;
 
-    idx = hash_djb2(key) % size;
-
-    return (idx);
+	idx = hash_djb2(key) % size;
+	return (idx);
 }
